// dist/tmp/app/shared/interfaces/commerce.interface.ts
export type PurchaseType =
  'SystemLicense' |
  'SystemLicenseNoDelivery' |
  'OfflineLicense' |
  'OfflineLicenseNoDelivery' |
  'PrepaySystemLicense' |
  'PrepayOfflineLicense' |
  'Trial' |
  'DeliveryOnly' |
  'ChannelNoDelivery' |
  'RevenueOnly' |
  'PrepayAccess';

export type PaymentOption =
  'Trial' |
  'DeliveryOnly' |
  'PurchaseOnCredit' |
  'CreditCard' |
  'Hold';

export type PaymentType =
  'PurchaseOnCredit' |
  'CreditCard';

export type QuoteStatus = 'ACTIVE' | 'PENDING' | 'ORDERED' | 'EXPIRED' | 'CANCELLED';

export type TranscodeStatus = 'Submitted' | 'Completed' | 'Failed' | 'UrlError' | 'Deleted';

export type OrderStatus = 'ORDER' | 'REFUND';

export type EditableQuoteFields = 'bulkOrderId' | 'discount' | 'purchaseType';

export const quotesWithoutPricing: PurchaseType[] = [
  'Trial',
  'DeliveryOnly'
];

export const quotesAllowedToHaveFeesOnly: PurchaseType[] = [
  'RevenueOnly',
  'PrepayAccess'
];

// Base interfaces

export interface CommonCommerce extends Common {
  projects?: Project[];
}

export interface Project {
  id: string;
  name: string;
  clientName: string;
  subTotal?: number;
  creditMemoForProjectId?: number;
  lineItems?: Array<AssetLineItem>;
  feeLineItems?: Array<FeeLineItem>;
  attributes?: Array<SelectedPriceAttribute>;
  [index: string]: any;
}

export interface LineItemNote {
  notes: string[];
}

export interface AssetLineItem {
  id?: string;
  asset?: Asset;
  notes?: Array<LineItemNote>;
  itemPrice?: number;
  multiplier?: number;
  transcodeStatus?: TranscodeStatus;
  downloadUrl?: string;
  asperaSpec?: string;
  tomSubmitTimestamp?: number;
  tomCompleteTimestamp?: number;
  tomTransactionId?: string;
  selectedTranscodeTarget?: string;
  transcodeTargets?: Array<string>;
  expirationDate?: number;
  attributes?: Array<SelectedPriceAttribute>;
  salesForceId?: string;
  price?: number;
  grossAssetPrice?: number;
  rightsManaged?: string;
  externalAgreementIds?: string[];
}

export interface FeeLineItem {
  id?: string;
  amount?: string;
  feeType?: string;
  notes?: string;
  salesForceId?: string;
}

export interface Asset {
  assetId?: number;
  assetName?: string;
  assetDuration?: number;
  metadata?: Metadatum[];
  rightsManaged?: string;
  supplierId?: number;
  supplierName?: string;
  thumbnailUrl?: string;
  timeStart?: number;
  timeEnd?: number;
  clipUrl?: string;
  uuid?: string;
  price?: number;
  masterDownloadUrl?: string;
}

export interface Metadatum {
  name: string;
  value: string;
}
// ...
export interface Quote extends CommonCommerce {
  createdUserId: number;
  ownerUserId: number;
  total: number;
  subTotal?: number;
  quoteStatus: QuoteStatus;
  purchaseType?: PurchaseType;
  projects?: Project[];
  itemCount?: number;
  expirationDate?: string;
  focused?: boolean;
  stripePublicKey?: string;
  bulkOrderId?: string;
  discount?: number;
  externalAgreementIds?: string[];
  internalAgreementIds?: number[];
  externalLicenseIds?: string[];
  internalLicenseIds?: number[];
  billingAccountId?: number;
  invoiceContact?: ViewAddress;
  billingAccountData?: SendDetailsBillingAccount;
  salesManager?: string;
  orderId?: number;
  ownerData?: {
    accountName: string;
    email: string;
    firstName: string;
    lastName: string;
  };
}

export interface QuoteEdit {
  data: Quote;
  sendDetails: SendDetails;
  loading: boolean;
}

export interface SendDetails {
  user: SendDetailsUser;
  billingAccount: SendDetailsBillingAccount;
  invoiceContact: SendDetailsInvoiceContact;
  salesManager: SendDetailsSalesManager;
}

export interface SendDetailsUser {
  id?: number;
  accountName?: string;
  customerName?: string;
  email?: string;
  [index: string]: any;
}

export interface SendDetailsBillingAccount {
  id?: number;
  name?: string;
  salesOwner?: string;
  salesManager?: string;
  purchaseOnCredit?: number;
  creditExemption?: number;
  paymentTermsDays?: string;
  readonly readonlyPaymentTermsDays?: string;
  licensingVertical?: string;
  invoiceContactId?: number;
  [index: string]: any;
}

export interface SendDetailsSalesManager {
  expirationDate?: string;
  salesManager?: string;
  offlineAgreement?: string;
  [index: string]: any;
}

export interface SendDetailsInvoiceContact {
  id?: number;
  name?: string;
  contactEmail?: string;
  contacts?: SendDetailsBillingAccount[];
  [index: string]: any;
}
// ...
export interface CommerceMessage {
  type: string;
  payload?: any;
}